generate purescript javascript; update index.html
We must generate the javascript with the following command:

pulp build -O -t public/js/main_ps.js

Those who read `pulp build -h` might notice that -O is optimise and try
to take it out of the command while developing to speed things up. If
you try this, you'll notice that no file is written. I've filed a bug
report for this:

https://github.com/bodil/pulp/issues/47

So if you have troubles with pulp not writing files, try adding the -O flag.